scene changing v0.5
Added demo scene changing. Currently you can approach a flagpole to go from a green rectangle to a grey rectangle! These are two different scenes ("home" and "school") that get added and deleted as a child to a "game" scene. The time and day are stored as variables in an autoloaded script "stats"